@@ -62,9 +62,9 @@
     "react-modal": "3.16.1",
     "react-motion": "0.5.2",
     "react-router-dom": "^5.1.2",
-    "react-tooltip": "4.2.21",
-    "reactotron-core-server": "*",
-    "reactotron-core-ui": "*",
+    "react-tooltip": "4.5.1",
+    "reactotron-core-server": "workspace:*",
+    "reactotron-core-ui": "workspace:*",
     "source-map-support": "^0.5.16",
     "styled-components": "^4.4.1"
   },
@@ -80,13 +80,13 @@
     "@storybook/react": "^5.2.8",
     "@testing-library/react-hooks": "^8.0.1",
     "@types/jest": "^29.2.5",
-    "@types/react": "^16.9.16",
-    "@types/react-dom": "^16.9.4",
+    "@types/react": "17.0.2",
+    "@types/react-dom": "17.0.2",
     "@types/react-router-dom": "^5.1.3",
     "@types/styled-components": "^4.4.1",
     "@typescript-eslint/eslint-plugin": "^2.12.0",
     "@typescript-eslint/parser": "^2.12.0",
-    "babel-loader": "^8.0.6",
+    "babel-loader": "^8.3.0",
     "electron": "22.0.1",
     "electron-builder": "^23.6.0",
     "electron-webpack": "2.8.2",
@@ -103,11 +103,11 @@
     "jest-environment-jsdom": "^29.3.1",
     "mime": "^3.0.0",
     "prettier": "^2.8.3",
-    "rimraf": "4.1.2",
     "react-test-renderer": "18.2.0",
+    "rimraf": "4.1.2",
     "ts-jest": "^29.0.4",
     "typescript": "4.9.5",
-    "webpack": "^4.41.3"
+    "webpack": "patch:webpack@4.46.0#../../.yarn/patches/webpack-npm-4.46.0-a8158cfa7f.patch"
   },
   "build": {
     "productName": "Reactotron",
