@@ -4,7 +4,7 @@ import { FaMagic } from "react-icons/fa"
 import styled from "styled-components"
 
 import SideBarButton from "../SideBarButton"
-const reactotronLogo = require("../../images/Reactotron-128.png")
+import { reactotronLogo } from "../../images"
 
 interface SideBarContainerProps {
   isOpen: boolean
@@ -13,11 +13,11 @@ const SideBarContainer = styled.div<SideBarContainerProps>`
   display: flex;
   flex-direction: column;
   padding-top: 25px;
-  background-color: ${props => props.theme.backgroundSubtleDark};
-  border-right: 1px solid ${props => props.theme.chromeLine};
+  background-color: ${(props) => props.theme.backgroundSubtleDark};
+  border-right: 1px solid ${(props) => props.theme.chromeLine};
   width: 115px;
   transition: margin 0.2s ease-out;
-  margin-left: ${props => (props.isOpen ? 0 : -115)}px;
+  margin-left: ${(props) => (props.isOpen ? 0 : -115)}px;
 `
 
 const Spacer = styled.div`
